Add files via upload
As a frequent user of the mobile app, I have come to appreciate the innovative and user-friendly system employed by the company. The app has made it incredibly easy for users to access multiple retail stores that would otherwise require a car trip that could be inconvenient, particularly depending on one’s location. This project analyzes data provided by instacart to provide KPIs for the company to make progressive decisions!